+++ buession-security-captcha/src/main/java/com/buession/security/captcha/utils/ResponseUtils.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2024 Buession.com Inc.														       |
+ * | Copyright @ 2013-2025 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.security.captcha.utils;
@@ -30,22 +30,23 @@ import java.io.IOException;
 
 /**
  * {@link Response} 工具类
+ *
  * @author Yong.Teng
  * @since 3.0.0
  */
 public class ResponseUtils {
 
-	protected ResponseUtils(){
+	protected ResponseUtils() {
 
 	}
 
-	public static void close(final Response response){
+	public static void close(final Response response) {
 		if(response != null){
 			if(response.getInputStream() != null){
 				try{
 					response.getInputStream().close();
 				}catch(IOException e){
-
+					//
 				}
 			}
 		}