+++ buession-security-pac4j/src/main/java/com/buession/security/pac4j/http/TextAjaxRequestResolver.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2025 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.security.pac4j.http;
@@ -27,14 +27,15 @@ package com.buession.security.pac4j.http;
 import com.buession.lang.Constants;
 import org.pac4j.core.context.HttpConstants;
 import org.pac4j.core.context.WebContext;
+import org.pac4j.core.context.session.SessionStore;
 import org.pac4j.core.exception.http.HttpAction;
 import org.pac4j.core.exception.http.RedirectionAction;
-import org.pac4j.core.exception.http.RedirectionActionHelper;
 import org.pac4j.core.exception.http.UnauthorizedAction;
 import org.pac4j.core.exception.http.WithLocationAction;
 import org.pac4j.core.http.ajax.DefaultAjaxRequestResolver;
 import org.pac4j.core.redirect.RedirectionActionBuilder;
 import org.pac4j.core.util.CommonHelper;
+import org.pac4j.core.util.HttpActionHelper;
 
 import java.util.Optional;
 
@@ -47,12 +48,13 @@ import java.util.Optional;
 public class TextAjaxRequestResolver extends DefaultAjaxRequestResolver {
 
 	@Override
-	public HttpAction buildAjaxResponse(final WebContext context,
-										final RedirectionActionBuilder redirectionActionBuilder){
+	public HttpAction buildAjaxResponse(final WebContext context, final SessionStore sessionStore,
+										final RedirectionActionBuilder redirectionActionBuilder) {
 		String url = null;
 
 		if(isAddRedirectionUrlAsHeader()){
-			final RedirectionAction action = redirectionActionBuilder.getRedirectionAction(context).orElse(null);
+			final RedirectionAction action = redirectionActionBuilder.getRedirectionAction(context, sessionStore)
+					.orElse(null);
 			if(action instanceof WithLocationAction){
 				url = ((WithLocationAction) action).getLocation();
 			}
@@ -62,10 +64,10 @@ public class TextAjaxRequestResolver extends DefaultAjaxRequestResolver {
 			if(CommonHelper.isNotBlank(url)){
 				context.setResponseHeader(HttpConstants.LOCATION_HEADER, url);
 			}
-			throw UnauthorizedAction.INSTANCE;
+			throw new UnauthorizedAction();
 		}
 
-		return RedirectionActionHelper.buildFormPostContentAction(context,
+		return HttpActionHelper.buildFormPostContentAction(context,
 				Optional.ofNullable(url).orElse(Constants.EMPTY_STRING));
 	}